doc: alternate function node example
I don't know if I made this example more intense than it needs to be, but, writing it out this way and running it helped my understanding. The existing example does not match the gif. Double spaced to squeeze the comments in. Let me know and I'll add the gif.